# service/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Studio 实验室管理系统 API",
        "title": "ScienceOL Studio API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "harveyque@outlook.com"
        },
        "license": {
            "name": "GNU Affero General Public License v3.0",
            "url": "http://www.gnu.org/licenses/agpl-3.0.en.html"
        },
        "version": "1.0"
    },
    "host": "localhost:48197",
    "basePath": "/api",
    "paths": {
        "/auth/callback/casdoor": {
            "get": {
                "description": "处理OAuth2授权回调",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "OAuth2回调",
                "parameters": [
                    {
                        "type": "string",
                        "description": "授权码",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "防CSRF攻击的状态码",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "重定向到前端错误页面",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "get": {
                "description": "检查服务运行状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "前端回调地址",
                        "name": "frontend_callback_url",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "重定向到OAuth2授权页面",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "重定向的授权URL地址"
                            }
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "刷新OAuth2令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "刷新令牌",
                "parameters": [
                    {
                        "description": "刷新令牌请求",
                        "name": "refresh_token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/login.RefreshTokenReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刷新 token 失败 code = 1002",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "code.ErrCode": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                1000,
                1001,
                1002,
                1003,
                5000,
                5001,
                5002,
                5003,
                5004,
                5005,
                5006,
                5007,
                5008,
                5009,
                5010,
                5011,
                5012,
                10000,
                10001,
                10002,
                10003,
                10004,
                10005,
                10006,
                10007,
                10008,
                10009,
                10010,
                20000,
                20001,
                20002,
                20003,
                20004,
                20005,
                20006,
                20007,
                20008,
                20009,
                20010,
                22000,
                22001,
                22002,
                22003,
                22004,
                22005,
                22006,
                22007,
                22008,
                22009,
                22010,
                22011,
                22012,
                22013,
                22014,
                22015,
                22016,
                22017,
                22018,
                22019,
                24000,
                24001,
                24002,
                26000,
                26001,
                26002,
                26003,
                26004,
                26005,
                28000,
                28001,
                28002,
                28003,
                28004,
                28005,
                28006,
                28007,
                28008,
                30000,
                30001,
                30002,
                30003,
                30004,
                30005,
                30006,
                30007,
                30008,
                30009,
                30010,
                30011,
                30012,
                30013,
                30014,
                30015,
                30016,
                30017,
                30018,
                30019,
                30020,
                30021,
                30022,
                30023,
                30024,
                30025,
                30026,
                30027,
                30028,
                30029,
                30030,
                30031,
                30032
            ],
            "x-enum-comments": {
                "BohrBatchQueryErr": "bhor batch query user error",
                "CallbackJobStatusKeyNotExistErr": "callback job status key note exists error",
                "CallbackJobStatusTimeoutErr": "callback job status timeout error",
                "CallbackParamErr": "callback parameter error",
                "CanNotFoundEdgeSession": "can not found edge session",
                "CanNotFoundMaterialNodeErr": "can not found material node error",
                "CanNotFoundTargetNode": "can not found target node error",
                "CanNotFoundWorkflowHandleErr": "can not found workflow handle error",
                "CanNotGetLabIDErr": "cannot get lab id error",
                "CanNotGetParentJobErr": "can not found parent node job error",
                "CanNotGetWorkflowTaskErr": "can not get workflow task error",
                "CanNotGetWorkflowUUIDErr": "can not get workflow uuid",
                "CanNotGetworkflowErr": "workflow not found error",
                "CasDoorCreateLabUserErr": "create lab user error",
                "CasDoorQueryLabUserErr": "query lab user error",
                "CreateDataErr": "database create data error",
                "DataNotMapAnyTypeErr": "data not map any type error",
                "DeleteDataErr": "database delete error",
                "EdgeConnectClosedErr": "connect closed when node running error",
                "EdgeHandleNotExistErr": "node handle not exist",
                "EdgeNodeNotExistErr": "edge node not exist",
                "ExchangeTokenErr": "exchange token failed",
                "ExecWorkflowNodeScriptErr": "exec workflow script error",
                "FormatCSVTaskErr": "format csv data error",
                "InvalidDagErr": "invalid dag error",
                "InvalidToken": "invalid token",
                "InvalidateJWT": "invalidate jwt",
                "InvalidateThirdID": "invalidate third id error",
                "InviteExpiredErr": "invite expired error",
                "JobCanceled": "job is canceled",
                "JobRetryTimeout": "job retry timeout error",
                "JobRunFailErr": "job run fail error",
                "JobTimeoutErr": "job timeout error",
                "LabAlreadyDeletedErr": "lab already deleted error",
                "LabIDIsEmptyErr": "lab id is empty error",
                "LabInviteNotFoundErr": "can not found laboratory invite link error",
                "LabNotFound": "laboratory not found error",
                "LoginCallbackErr": "login process user info failed",
                "LoginConfigErr": "login configuration error",
                "LoginFormatErr": "login verification format error",
                "LoginGetUserInfoErr": "get user info failed",
                "LoginSetStateErr": "set login state error",
                "LoginStateErr": "state verification failed",
                "MachineAlreadyExistErr": "machine already exist error",
                "MachineNodeStoppingErr": "machine is stopping",
                "MachineNotExistErr": "machine not exist error",
                "MachineReachMaxNumCountErr": "machine reach max number error",
                "MachineStartUnknownErr": "start machine unknown error",
                "MarshalTargetDataErr": "marshal target data error",
                "MaxTplNodeDeepErr": "max template node deep error",
                "ModelNotImplementTablerErr": "model not implement schema.Tabler",
                "NoPermission": "no permission",
                "NodeDataMarshalErr": "marshal node data error",
                "NotBaseDBTypeErr": "not base db type error",
                "NotPointerErr": "not pointer err",
                "NotSlicePointerErr": "must be a pointer to a slice",
                "NotifyActionAlreadyRegistryErr": "notify action already registry",
                "NotifySendMsgErr": "notify send message error",
                "NotifySubscribeChannelErr": "notify subscribe channel fail",
                "ParamDataKeyInvalidateErr": "param data key invalidate error",
                "ParamDataValueInvalidateErr": "param data value invalidate error",
                "ParamErr": "parse parameter error",
                "ParentNodeNotFoundErr": "parent node not found error",
                "ParseLoginRedirectURLErr": "redirect login url error",
                "PathHasEmptyName": "path has empty name error",
                "PermissionDenied": "permission denied",
                "PointerIsNilErr": "pointer is nil error",
                "QueryJobStatusKeyNotExistErr": "query job status key note exists error",
                "QueryMachineStatusFailErr": "query machine status error",
                "QueryRecordErr": "database query error",
                "RPCHttpCodeErr": "rpc request http code error",
                "RPCHttpCodeRespErr": "rpc request http code resp error",
                "RPCHttpErr": "rpc request http error",
                "RecordNotFound": "database record not found",
                "RedisAddSetErr": "redis add user set error",
                "RedisLuaRetErr": "redis lua return type error",
                "RedisLuaScriptErr": "redis lua script error",
                "RedisRemoveSetErr": "redis remove user set error",
                "RefreshTokenErr": "refresh token failed",
                "RefreshTokenParamErr": "refresh token parameter error",
                "RegActionNameEmptyErr": "reg action name empty",
                "ResNotExistErr": "resource not exist",
                "ResourceIsEmptyErr": "resource is empty",
                "ResourceNotExistErr": "resource not exist",
                "SaveWorkflowEdgeErr": "batch save workflow edge error",
                "SaveWorkflowNodeErr": "batch save nodes error",
                "Success": "success",
                "TargetDataNotMapAnyTypeErr": "target data not map any type error",
                "TargetParamInvalidateErr": "target param invalidate error",
                "TemplateNodeNotFoundErr": "template node not found error",
                "UnDefineErr": "undefined",
                "UnLogin": "not logged in",
                "UnknownWSActionErr": "unknown material websocket action",
                "UnknownWorkflowNodeTypeErr": "unknown workflow node type error",
                "UnmarshalWSDataErr": "unmarshal material websocket data error",
                "UpdateDataErr": "database update data error",
                "UpdateNodeErr": "update material node error",
                "UpsertWorkflowEdgeErr": "upsert workflow edge error",
                "UserIDIsEmptyErr": "user id is empty",
                "ValueNotExistErr": "value not exist error",
                "ValueSliceOutIndexErr": "value slice out index error",
                "WorkflowHasCircularErr": "workflow has circular error",
                "WorkflowNodeNoActionName": "workflow node no action name error",
                "WorkflowNodeNoActionType": "workflow node no action type error",
                "WorkflowNodeNoDeviceName": "workflow node no device name error",
                "WorkflowNodeNotFoundErr": "workflow node not found error",
                "WorkflowNodeScriptEmtpyErr": "workflow script empty error",
                "WorkflowNotExistErr": "workflow not exist",
                "WorkflowTaskAlreadyExistErr": "workflow task already exist error",
                "WorkflowTaskFinished": "workflow task finished",
                "WorkflowTaskNotFoundErr": "can not found workflow task error",
                "WorkflowTaskStatusErr": "workflow task status error",
                "WorkflowTaskStatusNotPendingErr": "workflow task not in pending status",
                "WorkflowTemplateNotFoundErr": "can not found workflow template error"
            },
            "x-enum-descriptions": [
                "success",
                "undefined",
                "no permission",
                "invalidate jwt",
                "parse parameter error",
                "not pointer err",
                "must be a pointer to a slice",
                "pointer is nil error",
                "login configuration error",
                "set login state error",
                "refresh token failed",
                "state verification failed",
                "exchange token failed",
                "callback parameter error",
                "get user info failed",
                "login process user info failed",
                "not logged in",
                "login verification format error",
                "invalid token",
                "refresh token parameter error",
                "redirect login url error",
                "database create data error",
                "database update data error",
                "database record not found",
                "database query error",
                "database delete error",
                "not base db type error",
                "model not implement schema.Tabler",
                "redis lua script error",
                "redis lua return type error",
                "redis add user set error",
                "redis remove user set error",
                "reg action name empty",
                "resource is empty",
                "resource not exist",
                "can not found workflow template error",
                "user id is empty",
                "lab id is empty error",
                "laboratory not found error",
                "can not found laboratory invite link error",
                "invite expired error",
                "invalidate third id error",
                "lab already deleted error",
                "resource not exist",
                "edge node not exist",
                "node handle not exist",
                "unknown material websocket action",
                "unmarshal material websocket data error",
                "cannot get lab id error",
                "update material node error",
                "parent node not found error",
                "template node not found error",
                "invalid dag error",
                "max template node deep error",
                "can not found material node error",
                "machine already exist error",
                "query machine status error",
                "machine not exist error",
                "machine reach max number error",
                "machine is stopping",
                "start machine unknown error",
                "can not found target node error",
                "path has empty name error",
                "notify action already registry",
                "notify subscribe channel fail",
                "notify send message error",
                "rpc request http error",
                "rpc request http code error",
                "rpc request http code resp error",
                "create lab user error",
                "query lab user error",
                "bhor batch query user error",
                "can not get workflow uuid",
                "workflow not exist",
                "upsert workflow edge error",
                "permission denied",
                "batch save nodes error",
                "batch save workflow edge error",
                "workflow node not found error",
                "workflow not found error",
                "format csv data error",
                "workflow task already exist error",
                "can not found edge session",
                "workflow has circular error",
                "connect closed when node running error",
                "marshal node data error",
                "job run fail error",
                "can not found workflow task error",
                "workflow task status error",
                "workflow task finished",
                "workflow node no device name error",
                "workflow node no action name error",
                "workflow node no action type error",
                "query job status key note exists error",
                "callback job status key note exists error",
                "job timeout error",
                "job retry timeout error",
                "callback job status timeout error",
                "job is canceled",
                "can not get workflow task error",
                "workflow task not in pending status",
                "can not found workflow handle error",
                "can not found parent node job error",
                "param data key invalidate error",
                "param data value invalidate error",
                "data not map any type error",
                "value slice out index error",
                "value not exist error",
                "target data not map any type error",
                "marshal target data error",
                "target param invalidate error",
                "workflow script empty error",
                "unknown workflow node type error",
                "exec workflow script error"
            ],
            "x-enum-varnames": [
                "Success",
                "UnDefineErr",
                "NoPermission",
                "InvalidateJWT",
                "ParamErr",
                "NotPointerErr",
                "NotSlicePointerErr",
                "PointerIsNilErr",
                "LoginConfigErr",
                "LoginSetStateErr",
                "RefreshTokenErr",
                "LoginStateErr",
                "ExchangeTokenErr",
                "CallbackParamErr",
                "LoginGetUserInfoErr",
                "LoginCallbackErr",
                "UnLogin",
                "LoginFormatErr",
                "InvalidToken",
                "RefreshTokenParamErr",
                "ParseLoginRedirectURLErr",
                "CreateDataErr",
                "UpdateDataErr",
                "RecordNotFound",
                "QueryRecordErr",
                "DeleteDataErr",
                "NotBaseDBTypeErr",
                "ModelNotImplementTablerErr",
                "RedisLuaScriptErr",
                "RedisLuaRetErr",
                "RedisAddSetErr",
                "RedisRemoveSetErr",
                "RegActionNameEmptyErr",
                "ResourceIsEmptyErr",
                "ResourceNotExistErr",
                "WorkflowTemplateNotFoundErr",
                "UserIDIsEmptyErr",
                "LabIDIsEmptyErr",
                "LabNotFound",
                "LabInviteNotFoundErr",
                "InviteExpiredErr",
                "InvalidateThirdID",
                "LabAlreadyDeletedErr",
                "ResNotExistErr",
                "EdgeNodeNotExistErr",
                "EdgeHandleNotExistErr",
                "UnknownWSActionErr",
                "UnmarshalWSDataErr",
                "CanNotGetLabIDErr",
                "UpdateNodeErr",
                "ParentNodeNotFoundErr",
                "TemplateNodeNotFoundErr",
                "InvalidDagErr",
                "MaxTplNodeDeepErr",
                "CanNotFoundMaterialNodeErr",
                "MachineAlreadyExistErr",
                "QueryMachineStatusFailErr",
                "MachineNotExistErr",
                "MachineReachMaxNumCountErr",
                "MachineNodeStoppingErr",
                "MachineStartUnknownErr",
                "CanNotFoundTargetNode",
                "PathHasEmptyName",
                "NotifyActionAlreadyRegistryErr",
                "NotifySubscribeChannelErr",
                "NotifySendMsgErr",
                "RPCHttpErr",
                "RPCHttpCodeErr",
                "RPCHttpCodeRespErr",
                "CasDoorCreateLabUserErr",
                "CasDoorQueryLabUserErr",
                "BohrBatchQueryErr",
                "CanNotGetWorkflowUUIDErr",
                "WorkflowNotExistErr",
                "UpsertWorkflowEdgeErr",
                "PermissionDenied",
                "SaveWorkflowNodeErr",
                "SaveWorkflowEdgeErr",
                "WorkflowNodeNotFoundErr",
                "CanNotGetworkflowErr",
                "FormatCSVTaskErr",
                "WorkflowTaskAlreadyExistErr",
                "CanNotFoundEdgeSession",
                "WorkflowHasCircularErr",
                "EdgeConnectClosedErr",
                "NodeDataMarshalErr",
                "JobRunFailErr",
                "WorkflowTaskNotFoundErr",
                "WorkflowTaskStatusErr",
                "WorkflowTaskFinished",
                "WorkflowNodeNoDeviceName",
                "WorkflowNodeNoActionName",
                "WorkflowNodeNoActionType",
                "QueryJobStatusKeyNotExistErr",
                "CallbackJobStatusKeyNotExistErr",
                "JobTimeoutErr",
                "JobRetryTimeout",
                "CallbackJobStatusTimeoutErr",
                "JobCanceled",
                "CanNotGetWorkflowTaskErr",
                "WorkflowTaskStatusNotPendingErr",
                "CanNotFoundWorkflowHandleErr",
                "CanNotGetParentJobErr",
                "ParamDataKeyInvalidateErr",
                "ParamDataValueInvalidateErr",
                "DataNotMapAnyTypeErr",
                "ValueSliceOutIndexErr",
                "ValueNotExistErr",
                "TargetDataNotMapAnyTypeErr",
                "MarshalTargetDataErr",
                "TargetParamInvalidateErr",
                "WorkflowNodeScriptEmtpyErr",
                "UnknownWorkflowNodeTypeErr",
                "ExecWorkflowNodeScriptErr"
            ]
        },
        "common.Error": {
            "type": "object",
            "properties": {
                "info": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "common.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/code.ErrCode"
                },
                "data": {},
                "error": {
                    "$ref": "#/definitions/common.Error"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
...
        "login.RefreshTokenReq": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}